requirement.txt
flask==3.0.2
python-pptx==0.6.23
python-dotenv
google-generativeai
PyPDF2==3.0.1